Generic DotCommonAttribute moved to a separate file
